requirements.txt
streamlit==1.45.1
Pillow==10.3.0
numpy==1.26.4
keras==3.6.0
tensorflow==2.18.0
scikit-learn==1.6.1